requirements.txt
django==4.1.4
django-rest-framework==3.14.0
uritemplate==4.1.1
